% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{chol_logdet}
\alias{chol_logdet}
\title{Cholesky log determinant}
\usage{
chol_logdet(C)
}
\arguments{
\item{C}{CHMfactor from \code{Matrix::Cholesky}}
}
\value{
numeric scalar
}
\description{
Computes \eqn{\log |A|} from a Cholesky factor.
}
