% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matern_sar.R
\name{.resolve_h}
\alias{.resolve_h}
\title{Resolve pixel size h from explicit value or domain_bbox + grid dims}
\usage{
.resolve_h(h, domain_bbox, m1, m2)
}
\description{
Resolve pixel size h from explicit value or domain_bbox + grid dims
}
\keyword{internal}
