% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{lanczos_logdet}
\alias{lanczos_logdet}
\title{Stochastic log determinant via Lanczos quadrature}
\usage{
lanczos_logdet(apply_A, probes, n_steps = 50L)
}
\arguments{
\item{apply_A}{function v -> Av (symmetric PD)}

\item{probes}{matrix of Rademacher probe vectors (p x n_probes)}

\item{n_steps}{number of Lanczos steps}
}
\description{
Stochastic log determinant via Lanczos quadrature
}
