% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{lanczos_quadrature}
\alias{lanczos_quadrature}
\title{Single Lanczos quadrature estimate for f(A) applied to probe vector
Uses Lanczos tridiagonalization + Gauss quadrature}
\usage{
lanczos_quadrature(apply_A, probe, f, n_steps)
}
\arguments{
\item{apply_A}{function v -> Av}

\item{probe}{probe vector}

\item{f}{scalar function to apply to eigenvalues}

\item{n_steps}{number of Lanczos steps}
}
\description{
Single Lanczos quadrature estimate for f(A) applied to probe vector
Uses Lanczos tridiagonalization + Gauss quadrature
}
