% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{pcg}
\alias{pcg}
\title{PCG solver
Solves A x = b iteratively via preconditioned conjugate gradient}
\usage{
pcg(apply_A, b, x0 = NULL, precond = NULL, tol = 1e-06, maxit = NULL)
}
\arguments{
\item{apply_A}{function v -> Av}

\item{b}{right hand side vector}

\item{x0}{initial guess, default zero}

\item{precond}{optional preconditioner function v -> M^{-1} v}

\item{tol}{convergence tolerance}

\item{maxit}{maximum iterations}
}
\description{
PCG solver
Solves A x = b iteratively via preconditioned conjugate gradient
}
