% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fastblm_fit}
\alias{predict.fastblm_fit}
\title{Predict from a fastblm_fit object}
\usage{
\method{predict}{fastblm_fit}(object, A_new, ...)
}
\arguments{
\item{object}{fastblm_fit object}

\item{A_new}{n_new x p design matrix for prediction}

\item{...}{unused}
}
\value{
numeric vector of predicted means
}
\description{
Predict from a fastblm_fit object
}
