% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_reml.R
\name{reml_project}
\alias{reml_project}
\title{REML projection}
\usage{
reml_project(y, A, X_fixed)
}
\arguments{
\item{y}{numeric response vector}

\item{A}{n x p design matrix}

\item{X_fixed}{n x q matrix of fixed covariates}
}
\value{
list with projected y, A, and n_eff
}
\description{
Projects y and A onto the orthogonal complement of X_fixed,
removing the influence of fixed effects on hyperparameter estimation.
}
