% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{lanczos_quadrature}
\alias{lanczos_quadrature}
\title{Single Lanczos quadrature estimate for f(A) applied to probe vector}
\usage{
lanczos_quadrature(apply_A, probe, f, n_steps)
}
\arguments{
\item{apply_A}{function \eqn{v \mapsto Av}}

\item{probe}{probe vector}

\item{f}{scalar function to apply to eigenvalues}

\item{n_steps}{number of Lanczos steps}
}
\value{
list with element \code{estimate}
}
\description{
Uses Lanczos tridiagonalization and Gauss quadrature to estimate
\eqn{v^\top f(A) v} for a scalar function \eqn{f}.
}
